refactor: split cli_utils.py into focused modules (#289)
Split the 637-line cli_utils.py into 6 focused modules following the
Single Responsibility Principle:

- errors.py: EmberCliError and error factory functions (103 lines)
- path_utils.py: Path normalization utilities (58 lines)
- editor.py: Editor integration for opening files (109 lines)
- progress.py: Progress bar and callback utilities (134 lines)
- result_lookup.py: Cache/hash lookup functions (138 lines)
- result_display.py: Content display with formatting (124 lines)

The original cli_utils.py is now a 87-line facade that re-exports
all symbols for backward compatibility. All modules are now under
200 lines as specified in the issue requirements.

Implements #289
